:recycle: chore : jpa 레포지토리 폴더 구분
