+++ src/main/java/project/airbnb/clone/controller/TourApiController.java
@@ -29,7 +29,7 @@ public class TourApiController {
         }
 	}
 	
-	@GetMapping("/api/accommodations")
+//	@GetMapping("/api/accommodations")
 	public List<Accommodation> getAllAccommodations() {
 	    return tourApiService.getAllAccommodations();
 	}
