+++ src/main/java/project/airbnb/clone/repository/jpa/AccommodationImageRepository.java
@@ -1,13 +1,16 @@
-package project.airbnb.clone.repository;
+package project.airbnb.clone.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import project.airbnb.clone.entity.Accommodation;
-import project.airbnb.clone.entity.AccommodationPrice;
+import project.airbnb.clone.entity.AccommodationImage;
 
 import java.util.List;
 
-public interface AccommodationPriceRepository extends JpaRepository<AccommodationPrice, Long> {
+public interface AccommodationImageRepository extends JpaRepository<AccommodationImage, Long> {
+
+    List<AccommodationImage> findByAccommodation(Accommodation accommodation);
+
     @Modifying(clearAutomatically = true)
     void deleteByAccommodationIn(List<Accommodation> accommodations);
 }
