+++ src/main/java/project/airbnb/clone/repository/jpa/AccommodationPriceRepository.java
@@ -1,16 +1,13 @@
-package project.airbnb.clone.repository;
+package project.airbnb.clone.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import project.airbnb.clone.entity.Accommodation;
-import project.airbnb.clone.entity.AccommodationAmenity;
+import project.airbnb.clone.entity.AccommodationPrice;
 
 import java.util.List;
 
-public interface AccommodationAmenityRepository extends JpaRepository<AccommodationAmenity, Long> {
-
-    List<AccommodationAmenity> findByAccommodation(Accommodation accommodation);
-
+public interface AccommodationPriceRepository extends JpaRepository<AccommodationPrice, Long> {
     @Modifying(clearAutomatically = true)
     void deleteByAccommodationIn(List<Accommodation> accommodations);
 }
