+++ src/main/java/project/airbnb/clone/repository/jpa/AmenityRepository.java
@@ -1,4 +1,4 @@
-package project.airbnb.clone.repository;
+package project.airbnb.clone.repository.jpa;
 
 import java.util.Collection;
 import java.util.List;
