+++ src/main/java/project/airbnb/clone/repository/jpa/ReservationRepository.java
@@ -1,4 +1,4 @@
-package project.airbnb.clone.repository;
+package project.airbnb.clone.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
