+++ src/main/java/project/airbnb/clone/service/TourApiService.java
@@ -14,7 +14,7 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 import project.airbnb.clone.entity.Accommodation;
-import project.airbnb.clone.repository.AccommodationRepository;
+import project.airbnb.clone.repository.jpa.AccommodationRepository;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
