+++ src/main/java/project/airbnb/clone/service/guest/GuestService.java
@@ -11,7 +11,7 @@ import project.airbnb.clone.consts.SocialType;
 import project.airbnb.clone.dto.guest.SignupRequestDto;
 import project.airbnb.clone.entity.Guest;
 import project.airbnb.clone.model.ProviderUser;
-import project.airbnb.clone.repository.guest.GuestRepository;
+import project.airbnb.clone.repository.jpa.GuestRepository;
 
 @Service
 @Transactional(readOnly = true)
