+++ src/test/java/project/airbnb/clone/service/jwt/TokenServiceTest.java
@@ -19,7 +19,7 @@ import project.airbnb.clone.common.jwt.JwtProperties.TokenProperties;
 import project.airbnb.clone.common.jwt.JwtProvider;
 import project.airbnb.clone.dto.jwt.TokenResponse;
 import project.airbnb.clone.entity.Guest;
-import project.airbnb.clone.repository.guest.GuestRepository;
+import project.airbnb.clone.repository.jpa.GuestRepository;
 import project.airbnb.clone.repository.redis.RedisRepository;
 
 import java.util.Base64;
